+++ CHANGELOG.md
@@ -1,7 +1,27 @@
 ## lifebit-ai/cloudos-cli: changelog
-## v2.28.0 (2025-06-13)
+
+## v2.31.0 (2025-06-17)
 
 - Adds command to show the path to logs and results of jobs
+
+## v2.30.0 (2025-06-16)
+
+### Feat
+
+- Unify all workflow import platforms into a single child class
+
+## v2.29.0 (2025-06-13)
+
+### Feat
+
+- Adds command to move files within and across project within the same workspace.
+
+## v2.28.0 (2025-06-11)
+
+### Feat
+
+- Adds new subcommand `details` for `job`, to retrieve and view job details either in stdout or json
+
 ## v2.27.0 (2025-06-10)
 
 ### Feat